몸박뎀, 피격시 무적시간

+++ Unity Project/Assets/Monster/Monster_stats.cs
@@ -77,7 +77,7 @@ public class Monster_stats : MonoBehaviour
         }
     }
 
-    public float give_damage()
+    public int give_damage()
     {
         return damage;
     }
